requirements.txt
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
delta-spark==3.3.1
exceptiongroup==1.3.0
executing==2.2.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
Pygments==2.19.1
pyspark==3.5.5
python-dateutil==2.9.0.post0
pyzmq==26.4.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
zipp==3.21.0